Refactor out template func into new theme.py file
